python-fastapi/requirements/docker.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
databases[asyncpg]==0.9.0
asyncpg==0.27.0
passlib[bcrypt]==1.7.4
google-auth-oauthlib==1.2.0
CacheControl==0.13.1
Jinja2==3.1.2
python-multipart==0.0.9